# docs/openapi.yaml
openapi: 3.1.0
info:
  title: SVM Scan API
  version: '1.0'
  description: |
    **SVM Scan** is a read-only REST API that serves indexed data for the
    Steem Virtual Machine chain — a Cosmos SDK chain with an integrated EVM.
    It's the backend for [svmscan.blazeapps.org](https://svmscan.blazeapps.org),
    but is itself a public, unauthenticated API that anyone can call directly.

    A background indexer walks the chain block-by-block into MongoDB
    (backfill from genesis, then a live tail of new blocks), and this API
    serves that indexed data. A few endpoints are deliberate exceptions that
    proxy the chain (or an external LCD/EVM RPC) live instead of reading
    from the index — each one says so explicitly in its description below.

    ## Conventions

    - **Pagination**: every list endpoint returns the same envelope —
      `{ "data": [...], "pagination": { "page": 0, "perPage": 50, "total": 123 } }`.
      `page` is 0-indexed. Pass `?page=` and `?perPage=` as query strings;
      omitting them uses the endpoint's own default page size.
    - **Errors**: non-2xx responses are `{ "error": "human-readable message" }`.
    - **Timestamps**: always ISO-8601 UTC strings (e.g. `2026-08-12T00:32:29.334Z`).
    - **Big numbers as strings**: coin amounts, block heights used as ids,
      registration/deposit/withdrawal ids, and anything else that can exceed
      JavaScript's safe integer range is serialized as a **string**, not a
...
  contact:
    name: SVM Scan
...
servers:
...
security: []
...
paths:
...
  /proposals:
    get:
      tags: [Proposals]
      operationId: listProposals
      summary: List governance proposals
      description: Newest first. Default page size 10.
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/perPage'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ProposalSummary'
                  pagination:
                    $ref: '#/components/schemas/PaginationMeta'
              example:
                data:
                  - id: 1
                    title: Set EVM min gas price to 0.0001 STEEM/gas
                    status: PROPOSAL_STATUS_VOTING_PERIOD
                    votingStartTime: '2026-08-21T10:12:43.401Z'
                    votingEndTime: '2026-08-23T10:12:43.401Z'
                    submitTime: '2026-08-21T10:12:43.401Z'
                pagination: { page: 0, perPage: 1, total: 7 }
...
  /proposals/{id}:
    get:
      tags: [Proposals]
      operationId: getProposal
      summary: Get a proposal
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
          example: 1
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProposalDetailResponse'
              example:
                id: 1
                title: Set EVM min gas price to 0.0001 STEEM/gas
                status: PROPOSAL_STATUS_VOTING_PERIOD
                votingStartTime: '2026-08-21T10:12:43.401Z'
                votingEndTime: '2026-08-23T10:12:43.401Z'
                submitTime: '2026-08-21T10:12:43.401Z'
                summary: Floors the EVM feemarket base fee at 100000000000000 asteem (0.0001 STEEM) per gas unit.
                proposer: steem1am0j95eum7tezy8n9fwjrj0nldkwaskg9xphh8
                depositEndTime: '2026-08-23T10:12:43.401Z'
                totalDeposit:
                  - { denom: asteem, amount: '10000000' }
                finalTallyResult:
                  yesCount: '100000000000000000000000'
                  noCount: '0'
                  abstainCount: '0'
                  noWithVetoCount: '0'
                expedited: false
                failedReason: ''
                messages:
                  - typeUrl: /cosmos.evm.feemarket.v1.MsgUpdateParams
                    data:
                      authority: steem10d07y265gmmuvt4z0w9aw880jnsr700jvhcqmf
                      params:
                        noBaseFee: false
                        baseFeeChangeDenominator: 8
                        elasticityMultiplier: 2
                        enableHeight: '0'
                        baseFee: '7.000000000000000000'
                        minGasPrice: '100000000000000.000000000000000000'
                        minGasMultiplier: '0.500000000000000000'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /bridge-deposits:
    get:
      tags: [Bridge Deposits]
      operationId: listBridgeDeposits
      summary: List bridge deposits
      description: Sorted by internal id descending (newest first). Default page size 25.
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/perPage'
        - name: status
          in: query
          schema: { type: string, enum: [pending, minted, unclaimable] }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/BridgeDeposit'
                  pagination:
                    $ref: '#/components/schemas/PaginationMeta'
              example:
                data:
                  - id: '1860'
                    txid: 66e253b62f033225894d2cd86236adb7b2880bb3
                    opIndex: 0
                    steemBlock: '108667185'
                    steemTimestamp: '2026-08-13T23:20:24Z'
                    steemSender: blaze.apps
                    gatewayAccount: svm.bank
                    amountMillisteem: '3'
                    memo: svm-deposit 0x7fb5a561b74e7da7cd5fb1497830e96b3310d857
                    derivedDestination: steem107662cdhfe760n2lk9yhsv8fdve3pkzhlmtsvu
                    destinationType: DESTINATION_TYPE_EVM
                    status: DEPOSIT_STATUS_MINTED
                    minted: true
                    mintedAt: '379199'
                    mintTxHash: 86f7f5541622f020b19d0d001adb0762439ab336edbd97ce6f3a8d7d4496c301
                    createdAtHeight: '379191'
                    validatorConfirmations:
                      - validatorAddress: steemvaloper1u96tympny5r8m76psg5jr7fm8kcrg76wms5jgt
                        moniker: dhaka.witness
                        timestamp: '1786663267'
                    asset: BRIDGE_ASSET_STEEM
                pagination: { page: 0, perPage: 1, total: 1743 }
...
  /bridge-deposits/{idOrTxid}:
    get:
      tags: [Bridge Deposits]
      operationId: getBridgeDeposit
      summary: Get a deposit by internal id or Steem txid
      parameters:
        - name: idOrTxid
          in: path
          required: true
          schema: { type: string }
          example: '8'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BridgeDeposit'
              example:
                id: '8'
                txid: 9be06037c04bcdfa61de6a701dbe2220e2f1ec1a
                opIndex: 0
                steemBlock: '107853469'
                steemTimestamp: '2026-07-16T14:05:06Z'
                steemSender: blaze.apps
                gatewayAccount: blaze.apps
                amountMillisteem: '318971'
                memo: svm-deposit steem107662cdhfe760n2lk9yhsv8fdve3pkzhlmtsvu
                derivedDestination: steem107662cdhfe760n2lk9yhsv8fdve3pkzhlmtsvu
                destinationType: DESTINATION_TYPE_COSMOS
                status: DEPOSIT_STATUS_MINTED
                minted: true
                mintedAt: '12693'
                mintTxHash: 1b901c472e1faa79393e826fc16ce3ccdcb07c14489e57beef797d5e6d93be16
                createdAtHeight: '12646'
                validatorConfirmations:
                  - validatorAddress: steemvaloper107662cdhfe760n2lk9yhsv8fdve3pkzh4gyrhl
                    moniker: blaze.apps
                    timestamp: '1784211183'
                asset: BRIDGE_ASSET_STEEM
        '404':
          $ref: '#/components/responses/NotFound'
  /bridge-withdrawals:
    get:
      tags: [Bridge Withdrawals]
      operationId: listBridgeWithdrawals
      summary: List bridge withdrawals
      description: Sorted by internal id descending. Default page size 25.
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/perPage'
        - name: status
          in: query
          schema: { type: string, enum: [requested, processed, refunded] }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/BridgeWithdrawal'
                  pagination:
                    $ref: '#/components/schemas/PaginationMeta'
              example:
                data:
                  - id: '1'
                    sender: steem1am0j95eum7tezy8n9fwjrj0nldkwaskg9xphh8
                    destinationSteemAccount: blaze.apps
                    amountAsteem: '1000000000000000000'
                    amountMillisteem: '998'
                    memo: test withdrawal
                    burnTxHash: 6625728b191751a5b488824f8141b7169e9e85f4d527a5e9c0f13a17a091a693
                    status: WITHDRAWAL_STATUS_PROCESSED
                    createdAtHeight: '10746'
                    asset: BRIDGE_ASSET_STEEM
                    feeMillisteem: '2'
                    steemPayoutTxid: efcf2883481218cc52ca111b731b743ac6f8ff91
                    payoutOpIndex: 0
                    processedAtHeight: '10767'
                    refundedAtHeight: '0'
                    validatorConfirmations:
                      - validatorAddress: steemvaloper1am0j95eum7tezy8n9fwjrj0nldkwaskg04wyvy
                        moniker: null
                        timestamp: '1787267155'
                pagination: { page: 0, perPage: 1, total: 2 }
...
  /bridge-withdrawals/{idOrBurnTxHash}:
    get:
      tags: [Bridge Withdrawals]
      operationId: getBridgeWithdrawal
      summary: Get a withdrawal by internal id or burn transaction hash
      parameters:
        - name: idOrBurnTxHash
          in: path
          required: true
          schema: { type: string }
          example: '2'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BridgeWithdrawal'
              example:
                id: '0'
                sender: steem1am0j95eum7tezy8n9fwjrj0nldkwaskg9xphh8
                destinationSteemAccount: blaze.apps
                amountAsteem: '1000000000000000000'
                amountMillisteem: '998'
                memo: test withdrawal
                burnTxHash: 5302a993cc263b147f41d46421be66fd1c94bb6cb05a0f895d692cbf9534c71e
                status: WITHDRAWAL_STATUS_PROCESSED
                createdAtHeight: '10732'
                asset: BRIDGE_ASSET_STEEM
                feeMillisteem: '2'
                steemPayoutTxid: 64ecd95b9a52e8557eedbeacb693d87e1c005e8e
                payoutOpIndex: 0
                processedAtHeight: '10764'
                refundedAtHeight: '0'
                validatorConfirmations:
                  - validatorAddress: steemvaloper1am0j95eum7tezy8n9fwjrj0nldkwaskg04wyvy
                    moniker: null
                    timestamp: '1787267140'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  parameters:
    page:
      name: page
      in: query
      description: 0-indexed page number.
      schema: { type: integer, default: 0, minimum: 0 }
    perPage:
      name: perPage
      in: query
      description: Items per page (endpoint-specific default and max).
      schema: { type: integer, minimum: 1, maximum: 200 }
...
  responses:
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example: { error: 'Not found' }
    BadRequest:
      description: Invalid parameter
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example: { error: 'Invalid block height' }
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
      required: [error]
    PaginationMeta:
      type: object
      properties:
        page: { type: integer }
        perPage: { type: integer }
        total: { type: integer }
      required: [page, perPage, total]
    Coin:
      type: object
      properties:
        denom: { type: string }
        amount: { type: string }
      required: [denom, amount]
...
    DecodeMsg:
      type: object
      description: A decoded Cosmos SDK message. `data`'s shape depends on `typeUrl` — see the message-type registry in `packages/shared/src/encoding/msg.ts`.
      properties:
        typeUrl: { type: string }
        data:
          type: [object, 'null']
          additionalProperties: true
      required: [typeUrl, data]
...
    ProposalSummary:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        status: { type: string }
        votingStartTime: { oneOf: [{ type: string, format: date-time }, { type: 'null' }] }
        votingEndTime: { oneOf: [{ type: string, format: date-time }, { type: 'null' }] }
        submitTime: { oneOf: [{ type: string, format: date-time }, { type: 'null' }] }
      required: [id, title, status, votingStartTime, votingEndTime, submitTime]
    ProposalDetailResponse:
      allOf:
        - $ref: '#/components/schemas/ProposalSummary'
        - type: object
          properties:
            summary: { type: string }
            proposer: { type: string }
            depositEndTime: { oneOf: [{ type: string, format: date-time }, { type: 'null' }] }
            totalDeposit:
              type: array
              items: { $ref: '#/components/schemas/Coin' }
            finalTallyResult:
              oneOf:
                - type: object
                  properties:
                    yesCount: { type: string }
                    noCount: { type: string }
                    abstainCount: { type: string }
                    noWithVetoCount: { type: string }
                  required: [yesCount, noCount, abstainCount, noWithVetoCount]
                - type: 'null'
            messages:
              type: array
              items: { $ref: '#/components/schemas/DecodeMsg' }
            expedited: { type: boolean, description: Gov v1's expedited-proposal path (shorter voting period, higher pass threshold). }
            failedReason: { type: string, description: 'Set only if status is PROPOSAL_STATUS_FAILED (passed, but its messages then failed to execute on-chain); empty string otherwise, including for PROPOSAL_STATUS_REJECTED.' }
          required: [summary, proposer, depositEndTime, totalDeposit, finalTallyResult, messages, expedited, failedReason]
...
    BridgeValidatorConfirmation:
      type: object
      properties:
        validatorAddress: { type: string }
        moniker: { oneOf: [{ type: string }, { type: 'null' }] }
        timestamp: { type: string, description: Unix seconds. }
      required: [validatorAddress, moniker, timestamp]
    BridgeDeposit:
      type: object
      properties:
        id: { type: string }
        txid: { type: string, description: Steem transaction id. }
        opIndex: { type: integer }
        steemBlock: { type: string }
        steemTimestamp: { type: string, format: date-time }
        steemSender: { type: string }
        gatewayAccount: { type: string }
        amountMillisteem: { type: string }
        memo: { type: string }
        derivedDestination: { type: string }
        destinationType: { type: string, enum: [DESTINATION_TYPE_NONE, DESTINATION_TYPE_COSMOS, DESTINATION_TYPE_EVM] }
        status: { type: string, enum: [DEPOSIT_STATUS_PENDING, DEPOSIT_STATUS_MINTED, DEPOSIT_STATUS_UNCLAIMABLE] }
        minted: { type: boolean }
        mintedAt: { type: string, description: 'SVM block height, "0" if not yet minted.' }
        mintTxHash: { type: string }
        createdAtHeight: { type: string }
        validatorConfirmations:
          type: array
          items: { $ref: '#/components/schemas/BridgeValidatorConfirmation' }
        asset: { type: string, enum: [BRIDGE_ASSET_STEEM, BRIDGE_ASSET_SBD], description: Both values are live on this chain — mints `asteem` or `asbd` respectively, both at 18 decimals under the same `amountMillisteem` numeric convention. Never assume STEEM; always read this field for the display unit. }
      required: [id, txid, opIndex, steemBlock, steemTimestamp, steemSender, gatewayAccount, amountMillisteem, memo, derivedDestination, destinationType, status, minted, mintedAt, mintTxHash, createdAtHeight, validatorConfirmations, asset]
...
    BridgeWithdrawal:
      type: object
      properties:
        id: { type: string }
        sender: { type: string, description: SVM account that burned the tokens. }
        destinationSteemAccount: { type: string }
        amountAsteem: { type: string, description: 18 decimals. }
        amountMillisteem: { type: string, description: Convenience 3-decimal duplicate of amountAsteem. }
        memo: { type: string }
        burnTxHash: { type: string, description: 'SVM chain tx hash (indexed — links to GET /transactions/{hash}).' }
        status: { type: string, enum: [WITHDRAWAL_STATUS_REQUESTED, WITHDRAWAL_STATUS_PROCESSED, WITHDRAWAL_STATUS_REFUNDED] }
        createdAtHeight: { type: string }
        asset: { type: string, enum: [BRIDGE_ASSET_STEEM, BRIDGE_ASSET_SBD], description: Both values are live on this chain — see BridgeDeposit's `asset` description. }
        feeMillisteem: { type: string, description: Protocol fee withheld from the withdrawal, sent to the bridge_reward module account. }
        steemPayoutTxid: { type: string, description: 'The Steem-side payout transaction id, set once status is PROCESSED. Empty string until then.' }
        payoutOpIndex: { type: integer }
        processedAtHeight: { type: string, description: 'SVM block height the withdrawal was marked PROCESSED at; "0" until then.' }
        refundedAtHeight: { type: string, description: 'SVM block height an unconfirmed withdrawal auto-refunded at past the timeout, without ever being attested; "0" unless that happened.' }
        validatorConfirmations:
          type: array
          items: { $ref: '#/components/schemas/BridgeValidatorConfirmation' }
      required: [id, sender, destinationSteemAccount, amountAsteem, amountMillisteem, memo, burnTxHash, status, createdAtHeight, asset, feeMillisteem, steemPayoutTxid, payoutOpIndex, processedAtHeight, refundedAtHeight, validatorConfirmations]
...